add condition to delete a task

+++ app/services/tasks/task_service.py
@@ -41,7 +41,7 @@ class TaskService:
     def delete_task_by_id(self, id_task):
         task = Task.query.filter(Task.task_id == id_task).first()
         result = False
-        if task:
+        if task and task.status.lower() == 'processed':
             db.session.delete(task)
             db.session.commit()
             original_file = File.query.filter(File.id == task.id_original_file).first()
@@ -49,7 +49,7 @@ class TaskService:
                 db.session.delete(original_file)
                 db.session.commit()
             result = True
-        return {'Deleted': result}
+        return result
     
 class orderby(Enum):
     desc = desc
